requirements.txt
streamlit
spacy
nltk
sentence-transformers
transformers
torch
networkx
tqdm
pinecone
python-dotenv
